camel_up/requirements.txt
pygame==2.5.2
mysql-connector-python==8.3.0
python-dotenv==1.0.1
